+++ app/src/main/java/org/breezyweather/sources/openmeteo/OpenMeteoService.kt
@@ -225,6 +225,7 @@ class OpenMeteoService @Inject constructor(
             ->
             convert(
                 context,
+                location,
                 openMeteoWeatherResult,
                 openMeteoAirQualityResult
             )