+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Version 5.1.9 (not yet released)
 
+**Weather sources**
+- [Open-Meteo] Fix issues related to DST
+
 **Improvements and fixes**
 - Fix current location keeping old timezone on some sources, when travelling to a different timezone
 - [Widgets] Tap on “Material You - Current” now works again